Refactored the fragments to activities and removed the navigation bar
