--- app/src/main/res/navigation/mobile_navigation.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/mobile_navigation"
-    app:startDestination="@+id/navigation_calender">
-
-    <fragment
-        android:id="@+id/navigation_home"
-        android:name="com.example.daytastic.ui.home.HomeFragment"
-        android:label="@string/title_home"
-        tools:layout="@layout/fragment_home" />
-
-    <fragment
-        android:id="@+id/navigation_calender"
-        android:name="com.example.daytastic.ui.calender.CalendarFragment"
-        android:label="@string/title_calender"
-        tools:layout="@layout/fragment_calender" />
-
-    <fragment
-        android:id="@+id/navigation_notifications"
-        android:name="com.example.daytastic.ui.notifications.NotificationsFragment"
-        android:label="@string/title_notifications"
-        tools:layout="@layout/fragment_notifications" />
-</navigation>
